// frontend/src/app/models/enums/category.enum.ts
export enum CategoryEnum {
  Jobs = 1,
  Ranks = 2,
  Segments = 3,
  Sectors = 4,
  BusinessGroups = 5,
  BusinessUnits = 6,
  FrontBackOffices = 7,
  Users = 8,
  Countries = 9,
  Locations = 10,
  Tracks = 11,
  Modules = 12,
  Badges = 13,
  PastEvents = 14,
  UserNameEmail = 19
}

export enum CareerEnum {
  College = 15,
  ComplementaryFormation = 16,
  ComplementaryInfo = 17,
  PerfilProseek = 18
}